added login security

+++ products/views.py
@@ -1,19 +1,23 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
 from .models import Product, Category
 from .forms import ProductForm
 
+# Create your views here.
+
 def all_products(request):
-    # Retrieve all products ordered by SKU
-    products = Product.objects.order_by('sku')
+    """ A view to show all products, including sorting and search queries """
+
+    products = Product.objects.all()
     query = None
     categories = None
     sort = None
     direction = None
-    
+
     if request.GET:
         if 'sort' in request.GET:
             sortkey = request.GET['sort']
@@ -28,15 +32,12 @@ def all_products(request):
                 if direction == 'desc':
                     sortkey = f'-{sortkey}'
             products = products.order_by(sortkey)
-
-    if request.GET:
+            
         if 'category' in request.GET:
             categories = request.GET['category'].split(',')
             products = products.filter(category__name__in=categories)
             categories = Category.objects.filter(name__in=categories)
 
-    
-    if request.GET:
         if 'q' in request.GET:
             query = request.GET['q']
             if not query:
@@ -54,22 +55,29 @@ def all_products(request):
         'current_categories': categories,
         'current_sorting': current_sorting,
     }
+
     return render(request, 'products/all_products.html', context)
+
 
 def product_detail(request, product_id):
     """ A view to show individual product details """
 
-    # Retrieve a specific product by its ID (product_id)
     product = get_object_or_404(Product, pk=product_id)
 
     context = {
-        'product': product,  
+        'product': product,
     }
 
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -85,32 +93,46 @@ def add_product(request):
     context = {
         'form': form,
     }
-    return render(request, 'products/add_product.html', context)
 
+    return render(request, template, context)
+
+
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
-    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
             form.save()
             messages.success(request, 'Successfully updated product!')
-            return redirect('product_detail', product_id=product.id)  # Redirect to the product detail page after successful form submission
+            return redirect(reverse('product_detail', args=[product.id]))
         else:
             messages.error(request, 'Failed to update product. Please ensure the form is valid.')
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.name}')
 
+    template = 'products/edit_product.html'
     context = {
         'form': form,
         'product': product,
     }
-    return render(request, 'products/edit_product.html', context)
 
+    return render(request, template, context)
+
+
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
